fix(security): deep audit batch -- domain separators, ct_eq, fail-closed boundaries, overflow, expect removal (bd-184k)
14 bugs fixed across 20 source files from CrimsonCrane security audit:

Domain separators:
- Add HMAC domain separator prefixes to trust_card (trust_card_registry_sig_v1)
  and epoch_scoped_keys (epoch_scoped_sign_v1) to prevent cross-context forgery

Constant-time comparison:
- Consolidate inline constant_time_eq implementations to centralized
  security::constant_time::ct_eq in divergence_gate, root_pointer, remote_cap,
  policy_checkpoint, trust_fabric
- Harden ct_eq_bytes: remove early return on unequal lengths (leaked length
  info), now iterates max(a.len(), b.len()) and XOR-accumulates mismatches

Fail-closed boundaries:
- Fix off-by-one in revocation_freshness, revocation_freshness_gate,
  proof_verifier: change <= to < so values at exact boundary are treated as
  expired (fail-closed policy)
- Fix supervision: change < to <= so restart at exact window boundary
  counts toward budget

Overflow/truncation:
- incident_bundle_retention: replace bare as u8 cast on u128 with
  pct.min(100) + u8::try_from().unwrap_or(100)
- control_lane_policy: use u16 intermediate for budget sum to prevent u8 wrap

Expect removal:
- Replace .expect() panics with Result propagation in universal_verifier_sdk,
  verifier_sdk, dpor_exploration, eviction_saga, isolation_mesh,
  approval_workflow, divergence_gate, root_pointer (new SigningKeyInvalid
  error variant), remote_cap (new CryptoEngineUnavailable variant)
- root_pointer: add stale-epoch lower-bound check in bootstrap_root

+++ crates/franken-node/src/control_plane/control_lane_policy.rs
@@ -424,7 +424,7 @@ impl ControlLanePolicy {
 
     /// Check that budgets sum to 100% (INV-CLP-BUDGET-SUM).
     pub fn verify_budget_sum(&self) -> bool {
-        let total: u8 = self.budgets.values().map(|b| b.min_pct).sum();
+        let total: u16 = self.budgets.values().map(|b| b.min_pct as u16).sum();
         total == 100
     }
